+++ src/v5/client/connector.rs
@@ -24,8 +24,8 @@ pub struct MqttConnector<A, T> {
     address: A,
     connector: T,
     pkt: codec::Connect,
-    handshake_timeout: u64,
-    disconnect_timeout: u64,
+    handshake_timeout: u16,
+    disconnect_timeout: u16,
 }
 
 impl<A> MqttConnector<A, ()>
@@ -159,8 +159,8 @@ where
     ///
     /// Handshake includes `connect` packet and response `connect-ack`.
     /// By default handshake timeuot is disabled.
-    pub fn handshake_timeout(mut self, timeout: usize) -> Self {
-        self.handshake_timeout = timeout as u64;
+    pub fn handshake_timeout(mut self, timeout: u16) -> Self {
+        self.handshake_timeout = timeout as u16;
         self
     }
 
@@ -172,8 +172,8 @@ where
     /// To disable timeout set value to 0.
     ///
     /// By default disconnect timeout is set to 3 seconds.
-    pub fn disconnect_timeout(mut self, timeout: usize) -> Self {
-        self.disconnect_timeout = timeout as u64;
+    pub fn disconnect_timeout(mut self, timeout: u16) -> Self {
+        self.disconnect_timeout = timeout as u16;
         self
     }
 
@@ -223,7 +223,7 @@ where
         if self.handshake_timeout > 0 {
             Either::Left(
                 Select::new(
-                    delay_for(Duration::from_millis(self.handshake_timeout)),
+                    delay_for(Duration::from_millis(self.handshake_timeout as u64)),
                     self._connect(),
                 )
                 .map(|result| match result {
@@ -270,7 +270,7 @@ where
                             state.with_codec(|codec| codec.set_max_outbound_size(size));
                         }
                         // server keep-alive
-                        let keep_alive = pkt.server_keepalive_sec.unwrap_or(keep_alive) as u64;
+                        let keep_alive = pkt.server_keepalive_sec.unwrap_or(keep_alive);
 
                         Ok(Client::new(
                             io,
