+++ src/v5/server.rs
@@ -27,8 +27,8 @@ pub struct MqttServer<Io, St, C: ServiceFactory, Cn: ServiceFactory, P: ServiceF
     max_size: u32,
     max_receive: u16,
     max_qos: Option<QoS>,
-    handshake_timeout: usize,
-    disconnect_timeout: usize,
+    handshake_timeout: u16,
+    disconnect_timeout: u16,
     max_topic_alias: u16,
     pool: Rc<MqttSinkPool>,
     _t: PhantomData<(Io, St)>,
@@ -87,7 +87,7 @@ where
     ///
     /// Handshake includes `connect` packet and response `connect-ack`.
     /// By default handshake timeuot is disabled.
-    pub fn handshake_timeout(mut self, timeout: usize) -> Self {
+    pub fn handshake_timeout(mut self, timeout: u16) -> Self {
         self.handshake_timeout = timeout;
         self
     }
@@ -100,7 +100,7 @@ where
     /// To disable timeout set value to 0.
     ///
     /// By default disconnect timeout is set to 3 seconds.
-    pub fn disconnect_timeout(mut self, val: usize) -> Self {
+    pub fn disconnect_timeout(mut self, val: u16) -> Self {
         self.disconnect_timeout = val;
         self
     }
@@ -277,7 +277,7 @@ fn handshake_service_factory<Io, St, C>(
     max_receive: u16,
     max_topic_alias: u16,
     max_qos: Option<QoS>,
-    handshake_timeout: usize,
+    handshake_timeout: u16,
     pool: Rc<MqttSinkPool>,
 ) -> impl ServiceFactory<
     Config = (),
@@ -325,7 +325,7 @@ fn handshake_service_factory2<Io, St, C>(
     max_receive: u16,
     max_topic_alias: u16,
     max_qos: Option<QoS>,
-    handshake_timeout: usize,
+    handshake_timeout: u16,
     pool: Rc<MqttSinkPool>,
 ) -> impl ServiceFactory<
     Config = (),
